refactor: extract utils and PiloteUIMessage
