+++ src/client/components/_commons/ChatUI/utils.ts
@@ -1,0 +1,13 @@
+import { PiloteUIMessage } from "@/server/albert/PiloteUIMessage";
+
+export const extractMessageText = (message: PiloteUIMessage): string => {
+  if (!message.parts) return "";
+  return message.parts
+    .map((part) => {
+      if (part.type === "text") {
+        return part.text;
+      }
+      return "";
+    })
+    .join("");
+};
